Bug #853755 [SC9863-1][sharkl3][8.1][User][CTS] 在CtsCameraTestCases中， android.hardware.cts.CameraTest#testPreviewPictureSizesCombination - failed.
[root cause  ] During capture error is received but there is no pending
capture request. Further configure stream comes, stops preview/capture.
In parallel error recovery start with no preview and only snapshot is
enable. This enables full path but binning path is disabled. Further
request for preview is received but its request cant be satisfied as
binning path is not enabled.
[changes     ] avoid channel start during recovery in case of preview
is not enabled and capture is idle (stopped by configure stream).
[side effects] no
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:no
[board]:s9863a_1h10
[test case]:preview/snapshot/video, cts.
[reviewers   ] miller.song
[change_type ] feature_bugfix
[tag_product ] specific

Change-Id: If17726935fccdf85142278cebb429a369c5b4f73